+++ pybop/_problem.py
@@ -13,6 +13,8 @@ class BaseProblem:
         The model to be used for the problem (default: None).
     check_model : bool, optional
         Flag to indicate if the model should be checked (default: True).
+    signal: List[str]
+      The signal to observe.
     init_soc : float, optional
         Initial state of charge (default: None).
     x0 : np.ndarray, optional
@@ -79,7 +81,8 @@ class BaseProblem:
 
     def evaluateS1(self, x):
         """
-        Evaluate the model with the given parameters and return the signal and its derivatives.
+        Evaluate the model with the given parameters and return the signal and
+        its derivatives.
 
         Parameters
         ----------
@@ -128,8 +131,8 @@ class FittingProblem(BaseProblem):
         The model to fit.
     parameters : list
         List of parameters for the problem.
-    dataset : list
-        List of data objects to fit the model to.
+    dataset : Dataset
+        Dataset object containing the data to fit the model to.
     signal : str, optional
         The signal to fit (default: "Voltage [V]").
     """
@@ -151,7 +154,7 @@ class FittingProblem(BaseProblem):
         # Check that the dataset contains time and current
         for name in ["Time [s]", "Current function [A]"] + self.signal:
             if name not in self._dataset:
-                raise ValueError(f"expected {name} in list of dataset")
+                raise ValueError(f"Expected {name} in list of dataset")
 
         self._time_data = self._dataset["Time [s]"]
         self.n_time_data = len(self._time_data)
@@ -160,14 +163,13 @@ class FittingProblem(BaseProblem):
         if np.any(self._time_data[:-1] >= self._time_data[1:]):
             raise ValueError("Times must be increasing.")
 
+        for signal in self.signal:
+            if len(self._dataset[signal]) != self.n_time_data:
+                raise ValueError(
+                    f"Time data and {signal} data must be the same length."
+                )
         target = [self._dataset[signal] for signal in self.signal]
         self._target = np.vstack(target).T
-        if self.n_outputs == 1:
-            if len(self._target) != self.n_time_data:
-                raise ValueError("Time data and target data must be the same length.")
-        else:
-            if self._target.shape != (self.n_time_data, self.n_outputs):
-                raise ValueError("Time data and target data must be the same shape.")
 
         # Add useful parameters to model
         if model is not None:
@@ -199,8 +201,12 @@ class FittingProblem(BaseProblem):
         ----------
         x : np.ndarray
             Parameter values to evaluate the model at.
-        """
 
+        Returns
+        -------
+        y : np.ndarray
+            The model output y(t) simulated with inputs x.
+        """
         if (x != self.x).all() and self._model.matched_parameters:
             for i, param in enumerate(self.parameters):
                 param.update(value=x[i])
@@ -220,6 +226,12 @@ class FittingProblem(BaseProblem):
         ----------
         x : np.ndarray
             Parameter values to evaluate the model at.
+
+        Returns
+        -------
+        tuple
+            A tuple containing the simulation result y(t) and the sensitivities dy/dx(t) evaluated
+            with given inputs x.
         """
         if self._model.matched_parameters:
             raise RuntimeError(
@@ -292,6 +304,11 @@ class DesignProblem(BaseProblem):
         ----------
         x : np.ndarray
             Parameter values to evaluate the model at.
+
+        Returns
+        -------
+        y : np.ndarray
+            The model output y(t) simulated with inputs x.
         """
 
         sol = self._model.predict(